Separation collider creation from the rigid body creation
