+++ rapier/src/body.rs
@@ -1,5 +1,4 @@
 use bevy::ecs::prelude::*;
-use bevy::math::prelude::*;
 use bevy::transform::prelude::*;
 use fnv::FnvHashMap;
 
@@ -9,51 +8,29 @@ use crate::convert::{IntoBevy, IntoRapier};
 use crate::rapier::dynamics::{
     BodyStatus, JointSet, RigidBodyBuilder, RigidBodyHandle, RigidBodySet,
 };
-use crate::rapier::geometry::{Collider, ColliderBuilder, ColliderHandle, ColliderSet};
-use crate::rapier::math::Point;
+use crate::rapier::geometry::{ColliderHandle, ColliderSet};
 
 pub(crate) type HandleMap = FnvHashMap<Entity, RigidBodyHandle>;
-
-trait ColliderFactory {
-    fn collider_builder(&self) -> ColliderBuilder;
-
-    fn build(&self, entity: Entity, body_type: RigidBody, material: PhysicMaterial) -> Collider {
-        let mut collider_builder = self.collider_builder();
-        collider_builder = collider_builder
-            .user_data(entity.to_bits().into())
-            .sensor(matches!(body_type, RigidBody::Sensor))
-            .restitution(material.restitution)
-            .density(material.density)
-            .friction(material.friction);
-        collider_builder.build()
-    }
-}
 
 #[allow(clippy::type_complexity)]
 pub(crate) fn create(
     mut commands: Commands<'_>,
     mut bodies: ResMut<'_, RigidBodySet>,
-    mut colliders: ResMut<'_, ColliderSet>,
     mut handles: ResMut<'_, HandleMap>,
     query: Query<
         '_,
         (
             Entity,
-            &CollisionShape,
             &GlobalTransform,
             Option<&RigidBody>,
             Option<&Velocity>,
-            Option<&PhysicMaterial>,
             Option<&RotationConstraints>,
         ),
         Without<RigidBodyHandle>,
     >,
 ) {
-    for (entity, body, transform, body_type, velocity, material, rotation_constraints) in
-        query.iter()
-    {
-        let body_type = body_type.copied().unwrap_or_default();
-
+    for (entity, transform, body, velocity, rotation_constraints) in query.iter() {
+        let body_type = body.copied().unwrap_or_default();
         let mut builder = RigidBodyBuilder::new(body_status(body_type))
             .user_data(entity.to_bits().into())
             .position((transform.translation, transform.rotation).into_rapier());
@@ -89,14 +66,9 @@ pub(crate) fn create(
         }
 
         let rigid_body_handle = bodies.insert(builder.build());
-        let collider = body.build(entity, body_type, material.copied().unwrap_or_default());
 
-        let collider_handle = colliders.insert(collider, rigid_body_handle, &mut bodies);
         handles.insert(entity, rigid_body_handle);
-        commands
-            .entity(entity)
-            .insert(rigid_body_handle)
-            .insert(collider_handle);
+        commands.entity(entity).insert(rigid_body_handle);
     }
 }
 
@@ -249,108 +221,10 @@ pub(crate) fn update_bevy_transform(
     }
 }
 
-impl ColliderFactory for CollisionShape {
-    fn collider_builder(&self) -> ColliderBuilder {
-        match self {
-            CollisionShape::Sphere { radius } => ColliderBuilder::ball(*radius),
-            CollisionShape::Capsule {
-                half_segment: half_height,
-                radius,
-            } => ColliderBuilder::capsule_y(*half_height, *radius),
-            CollisionShape::Cuboid { half_extends } => cuboid_builder(*half_extends),
-            CollisionShape::ConvexHull { points } => convex_hull_builder(points.as_slice()),
-        }
-    }
-}
-
-#[inline]
-#[cfg(feature = "2d")]
-fn cuboid_builder(half_extends: Vec3) -> ColliderBuilder {
-    ColliderBuilder::cuboid(half_extends.x, half_extends.y)
-}
-
-#[inline]
-#[cfg(feature = "3d")]
-fn cuboid_builder(half_extends: Vec3) -> ColliderBuilder {
-    ColliderBuilder::cuboid(half_extends.x, half_extends.y, half_extends.z)
-}
-
-#[inline]
-fn convex_hull_builder(points: &[Vec3]) -> ColliderBuilder {
-    let points: Vec<Point<f32>> = points.into_rapier();
-    ColliderBuilder::convex_hull(points.as_slice()).expect("Failed to create convex-hull")
-}
-
 fn body_status(body_type: RigidBody) -> BodyStatus {
     match body_type {
         RigidBody::Dynamic => BodyStatus::Dynamic,
         RigidBody::Static | RigidBody::Sensor => BodyStatus::Static,
         RigidBody::Kinematic => BodyStatus::Kinematic,
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use bevy::math::Vec3;
-
-    use super::*;
-
-    #[test]
-    fn build_sphere() {
-        let collider = CollisionShape::Sphere { radius: 4.2 }.build(
-            Entity::new(0),
-            RigidBody::default(),
-            Default::default(),
-        );
-
-        let ball = collider
-            .shape()
-            .as_ball()
-            .expect("Created shape was not a ball");
-        assert_eq!(ball.radius, 4.2);
-    }
-
-    #[test]
-    fn build_cuboid() {
-        let collider = CollisionShape::Cuboid {
-            half_extends: Vec3::new(1.0, 2.0, 3.0),
-        }
-        .build(Entity::new(0), RigidBody::default(), Default::default());
-
-        let cuboid = collider
-            .shape()
-            .as_cuboid()
-            .expect("Created shape was not a cuboid");
-
-        assert_eq!(cuboid.half_extents.x, 1.0);
-        assert_eq!(cuboid.half_extents.y, 2.0);
-
-        #[cfg(feature = "3d")]
-        assert_eq!(cuboid.half_extents.z, 3.0);
-    }
-
-    #[test]
-    fn build_capsule() {
-        let collider = CollisionShape::Capsule {
-            half_segment: 10.0,
-            radius: 5.0,
-        }
-        .build(Entity::new(0), RigidBody::default(), Default::default());
-
-        let capsule = collider
-            .shape()
-            .as_capsule()
-            .expect("Created shape was not a capsule");
-
-        assert_eq!(capsule.radius, 5.0);
-        assert_eq!(capsule.segment.a.x, 0.0);
-        assert_eq!(capsule.segment.b.x, 0.0);
-        assert_eq!(capsule.segment.a.y, -10.0);
-        assert_eq!(capsule.segment.b.y, 10.0);
-
-        #[cfg(feature = "3d")]
-        assert_eq!(capsule.segment.a.z, 0.0);
-        #[cfg(feature = "3d")]
-        assert_eq!(capsule.segment.b.z, 0.0);
     }
 }
